move: 파일 정리
