--- src/components/Layout/FloatingButton.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { Plus } from "@/assets/icons";
-
-const FloatingButton = () => {
-  return (
-    <div className={"main-contain fixed bottom-0 w-full"}>
-      {/* TODO - 모집글로 이동 Link */}
-      <div className="absolute bottom-75 right-10 flex h-50 w-50 items-center justify-center rounded-full bg-black">
-        <Plus />
-      </div>
-    </div>
-  );
-};
-
-export default FloatingButton;
