--- src/components/Layout/Header.tsx
@@ -1,41 +1,0 @@
-"use client";
-
-import Modal from "../Common/Modal";
-import CategoryTab from "../Home/FilterPanel/CategoryTab";
-import NotificationSheet from "../Home/Notification/NotificationSheet";
-import { HIDDEN_PATH } from "@/assets/constantData";
-import { Bell } from "@/assets/icons";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import { useState } from "react";
-
-const Header = () => {
-  const path = usePathname();
-  const [isAlertOpen, setIsAlertOpen] = useState<boolean>(false);
-
-  return (
-    <>
-      <header className={`sticky top-0 z-10 flex flex-col bg-white ${HIDDEN_PATH.includes(path) && "hidden"}`}>
-        <div className="flex justify-between px-20 py-9">
-          <Link
-            href="/"
-            className="text-24 font-black leading-tight"
-          >
-            LOGO
-          </Link>
-          <button onClick={() => setIsAlertOpen(true)}>
-            <Bell />
-          </button>
-        </div>
-        <CategoryTab />
-      </header>
-      {isAlertOpen && (
-        <Modal>
-          <NotificationSheet setIsAlertOpen={setIsAlertOpen} />
-        </Modal>
-      )}
-    </>
-  );
-};
-
-export default Header;
