--- src/components/PostDetail/post/PostBody.tsx
@@ -1,5 +1,0 @@
-const PostBody = ({ content }) => {
-  return <div>{content}</div>;
-};
-
-export default PostBody;
