--- src/components/PostDetail/post/PostTitle.tsx
@@ -1,5 +1,0 @@
-const PostTitle = () => {
-  return <div>PostTitle</div>;
-};
-
-export default PostTitle;
