--- src/components/UI/CardLabel.tsx
@@ -1,15 +1,0 @@
-interface CardLabelProps {
-  title: string;
-  value: string;
-}
-
-const CardLabel = ({ title, value }: CardLabelProps) => {
-  return (
-    <>
-      <p className="caption-r text-black-400">{title}</p>
-      <p className="caption-r">{value}</p>
-    </>
-  );
-};
-
-export default CardLabel;
