--- src/components/UI/CardProfile.tsx
@@ -1,18 +1,0 @@
-interface CardProfileProps {
-  userNickname?: string;
-  modifiedAt?: string;
-}
-
-const CardProfile = ({ userNickname, modifiedAt }: CardProfileProps) => {
-  return (
-    <div className="flex items-center gap-x-12">
-      <div className="h-40 w-40 rounded-full bg-gray-300" />
-      <div className="flex flex-col gap-y-2">
-        <p className="body-14-m">{userNickname}</p>
-        <p className="caption-r">데이터 안 넘어옴</p>
-      </div>
-    </div>
-  );
-};
-
-export default CardProfile;
